WrapUpを一箇所へ集約
