+++ SuperNewRoles/EndGame/EndGame.cs
@@ -866,58 +866,6 @@ namespace SuperNewRoles.EndGame
             }
         }
     }
-    [HarmonyPatch(typeof(ExileController), nameof(ExileController.WrapUp))]
-    public class CheckEndGamePatch
-    {
-        public static void Prefix(ExileController __instance)
-        {
-            try
-            {
-                WrapUpClass.Prefix(__instance.exiled);
-            }
-            catch (Exception e)
-            {
-                SuperNewRolesPlugin.Logger.LogInfo("CHECKERROR:" + e);
-            }
-        }
-        public static void Postfix(ExileController __instance)
-        {
-            try
-            {
-                WrapUpClass.WrapUpPostfix(__instance.exiled);
-            }
-            catch (Exception e)
-            {
-                SuperNewRolesPlugin.Logger.LogInfo("CHECKERROR:" + e);
-            }
-        }
-    }
-    [HarmonyPatch(typeof(AirshipExileController), nameof(AirshipExileController.WrapUpAndSpawn))]
-    public class CheckAirShipEndGamePatch
-    {
-        public static void Prefix(AirshipExileController __instance)
-        {
-            try
-            {
-                WrapUpClass.Prefix(__instance.exiled);
-            }
-            catch (Exception e)
-            {
-                SuperNewRolesPlugin.Logger.LogInfo("CHECKERROR:" + e);
-            }
-        }
-        public static void Postfix(AirshipExileController __instance)
-        {
-            try
-            {
-                WrapUpClass.WrapUpPostfix(__instance.exiled);
-            }
-            catch (Exception e)
-            {
-                SuperNewRolesPlugin.Logger.LogInfo("CHECKERROR:"+e);
-            }
-        }
-    }
     [HarmonyPatch(typeof(TranslationController), nameof(TranslationController.GetString), new Type[] { typeof(StringNames), typeof(Il2CppReferenceArray<Il2CppSystem.Object>) })]
     class ExileControllerMessagePatch
     {
@@ -948,117 +896,8 @@ namespace SuperNewRoles.EndGame
         {
             PlayerControl.LocalPlayer.SetKillTimerUnchecked(RoleHelpers.GetEndMeetingKillCoolTime(PlayerControl.LocalPlayer),RoleHelpers.GetEndMeetingKillCoolTime(PlayerControl.LocalPlayer));
         }
-        public static void Prefix(GameData.PlayerInfo exiled)
-        {
-            RoleClass.IsCoolTimeSetted = false;
-            FalseCharges.WrapUp(exiled.Object);
-            if (!ModeHandler.isMode(ModeId.Default)) return;
-            if (ModeHandler.isMode(ModeId.Default))
-            {
-                if (SabotageManager.thisSabotage == SabotageManager.CustomSabotage.CognitiveDeficit){
-                    if (!Sabotage.CognitiveDeficit.main.IsLocalEnd)
-                    {
-                        Sabotage.CognitiveDeficit.main.UpdateTime = 0;
-                    }
-                }
-            }
-            if (exiled == null) return;
-            FinalStatusPatch.FinalStatusData.FinalStatuses[exiled.Object.PlayerId] = FinalStatus.Exiled;
-            if (exiled.Object.PlayerId != PlayerControl.LocalPlayer.PlayerId) return;
-            if (exiled.Object.isRole(RoleId.SideKiller))
-            {
-                var sideplayer = RoleClass.SideKiller.getSidePlayer(PlayerControl.LocalPlayer);
-                if (sideplayer != null)
-                {
-                    if (!RoleClass.SideKiller.IsUpMadKiller)
-                    {
-                        sideplayer.RPCSetRoleUnchecked(RoleTypes.Impostor);
-                        RoleClass.SideKiller.IsUpMadKiller = true;
-                    }
-                }
-            }
-        }
         public static void WrapUpPostfix(GameData.PlayerInfo exiled)
         {
-            SerialKiller.WrapUp();
-            PlayerControlHepler.refreshRoleDescription(PlayerControl.LocalPlayer);
-            new LateTask(() => {
-                RoleClass.IsMeeting = false;
-            }, 0.1f, "SetIsMeeting");
-            if (ModeHandler.isMode(ModeId.SuperHostRoles)) Mode.SuperHostRoles.WrapUpClass.WrapUp(exiled);
-            ModeHandler.Wrapup(exiled);
-            if (exiled == null) return;
-            exiled.Object.Exiled();
-            exiled.IsDead = true;
-            FinalStatusPatch.FinalStatusData.FinalStatuses[exiled.PlayerId] = FinalStatus.Exiled;
-            var Player = ModHelpers.playerById(exiled.PlayerId);
-            if (ModeHandler.isMode(ModeId.Default))
-            {
-                if (RoleClass.Lovers.SameDie && Player.IsLovers())
-                {
-                    if (AmongUsClient.Instance.AmHost)
-                    {
-                        PlayerControl SideLoverPlayer = Player.GetOneSideLovers();
-                        if (SideLoverPlayer.isAlive())
-                        {
-                            MessageWriter writer = AmongUsClient.Instance.StartRpcImmediately(PlayerControl.LocalPlayer.NetId, (byte)CustomRPC.CustomRPC.RPCMurderPlayer, Hazel.SendOption.Reliable, -1);
-                            writer.Write(SideLoverPlayer.PlayerId);
-                            writer.Write(SideLoverPlayer.PlayerId);
-                            writer.Write(byte.MaxValue);
-                            AmongUsClient.Instance.FinishRpcImmediately(writer);
-                            RPCProcedure.RPCMurderPlayer(SideLoverPlayer.PlayerId, SideLoverPlayer.PlayerId, byte.MaxValue);
-                        }
-                    }
-                }
-            }
-            if (ModeHandler.isMode(ModeId.Default))
-            {
-                EvilEraser.IsWinGodGuard = false;
-                EvilEraser.IsWinFoxGuard = false;
-                if (RoleHelpers.IsQuarreled(Player))
-                {
-                    var Side = RoleHelpers.GetOneSideQuarreled(Player);
-                    if (Side.isDead())
-                    {
-                        MessageWriter Writer = AmongUsClient.Instance.StartRpcImmediately(PlayerControl.LocalPlayer.NetId, (byte)CustomRPC.CustomRPC.ShareWinner, Hazel.SendOption.Reliable, -1);
-                        Writer.Write(Player.PlayerId);
-                        AmongUsClient.Instance.FinishRpcImmediately(Writer);
-                        CustomRPC.RPCProcedure.ShareWinner(Player.PlayerId);
-                        RoleClass.Quarreled.IsQuarreledWin = true;
-                        CheckGameEndPatch.CustomEndGame((GameOverReason)CustomGameOverReason.QuarreledWin, false);
-                    }
-                }
-
-                if (Roles.RoleClass.Jester.JesterPlayer.IsCheckListPlayerControl(Player))
-                {
-
-                    if (!Roles.RoleClass.Jester.IsJesterTaskClearWin || (Roles.RoleClass.Jester.IsJesterTaskClearWin && Patch.TaskCount.TaskDateNoClearCheck(Player.Data).Item2 - Patch.TaskCount.TaskDateNoClearCheck(Player.Data).Item1 == 0))
-                    {
-                        CustomRPC.RPCProcedure.ShareWinner(Player.PlayerId);
-
-                        MessageWriter Writer = AmongUsClient.Instance.StartRpcImmediately(PlayerControl.LocalPlayer.NetId, (byte)CustomRPC.CustomRPC.ShareWinner, Hazel.SendOption.Reliable, -1);
-                        Writer.Write(Player.PlayerId);
-                        AmongUsClient.Instance.FinishRpcImmediately(Writer);
-                        Roles.RoleClass.Jester.IsJesterWin = true;
-                        CheckGameEndPatch.CustomEndGame((GameOverReason)CustomGameOverReason.JesterWin, false);
-                    }
-                }
-
-                if (Roles.RoleClass.MadJester.MadJesterPlayer.IsCheckListPlayerControl(Player))
-                {
-
-                    if (!Roles.RoleClass.MadJester.IsMadJesterTaskClearWin || (Roles.RoleClass.MadJester.IsMadJesterTaskClearWin && Patch.TaskCount.TaskDateNoClearCheck(Player.Data).Item2 - Patch.TaskCount.TaskDateNoClearCheck(Player.Data).Item1 == 0))
-                    {
-                        CustomRPC.RPCProcedure.ShareWinner(Player.PlayerId);
-
-                        MessageWriter Writer = AmongUsClient.Instance.StartRpcImmediately(PlayerControl.LocalPlayer.NetId, (byte)CustomRPC.CustomRPC.ShareWinner, Hazel.SendOption.Reliable, -1);
-                        Writer.Write(Player.PlayerId);
-                        AmongUsClient.Instance.FinishRpcImmediately(Writer);
-                        Roles.RoleClass.MadJester.IsMadJesterWin = true;
-                        CheckGameEndPatch.CustomEndGame((GameOverReason)CustomGameOverReason.MadJesterWin, false);
-                    }
-                }
-            }
         }
     }
     [HarmonyPatch(typeof(ExileController), nameof(ExileController.ReEnableGameplay))]
