+++ SuperNewRoles/Roles/EvilGambler.cs
@@ -10,41 +10,6 @@ namespace SuperNewRoles.Roles
 {
     class EvilGambler
     {
-        [HarmonyPatch(typeof(ExileController), nameof(ExileController.WrapUp))]
-        public class CheckEndGamePatch
-        {
-            public static void Postfix(ExileController __instance)
-            {
-                try
-                {
-                    endm();
-                }
-                catch (Exception e)
-                {
-                    SuperNewRolesPlugin.Logger.LogInfo("EvilGambler:" + e);
-                }
-            }
-        }
-        [HarmonyPatch(typeof(AirshipExileController), nameof(AirshipExileController.WrapUpAndSpawn))]
-        public class CheckAirShipEndGamePatch
-        {
-            public static void Postfix(AirshipExileController __instance)
-            {
-                try
-                {
-                    endm();
-                }
-                catch (Exception e)
-                {
-                    SuperNewRolesPlugin.Logger.LogInfo("EvilGambler:" + e);
-                }
-            }
-        }
-        public static void endm() {
-            //HudManager.Instance.KillButton.SetCoolDown(EvilGamblerMurder.temp, EvilGamblerMurder.temp);
-            //PlayerControl.LocalPlayer.SetKillTimer(EvilGamblerMurder.temp);
-        }
-
         public static class EvilGamblerMurder
         {
             public static void Prefix(PlayerControl __instance, [HarmonyArgument(0)] PlayerControl target)
