+++ SuperNewRoles/Roles/Seer.cs
@@ -33,25 +33,8 @@ namespace SuperNewRoles.Roles
 
         public static class ExileControllerWrapUpPatch
         {
-            [HarmonyPatch(typeof(ExileController), nameof(ExileController.WrapUp))]
-            class BaseExileControllerPatch
-            {
-                public static void Postfix(ExileController __instance)
-                {
-                    WrapUpPostfix(__instance.exiled);
-                }
-            }
 
-            [HarmonyPatch(typeof(AirshipExileController), nameof(AirshipExileController.WrapUpAndSpawn))]
-            class AirshipExileControllerPatch
-            {
-                public static void Postfix(AirshipExileController __instance)
-                {
-                    WrapUpPostfix(__instance.exiled);
-                }
-            }
-
-            static void WrapUpPostfix(GameData.PlayerInfo exiled)
+            public static void WrapUpPostfix(GameData.PlayerInfo exiled)
             {
                 var role = PlayerControl.LocalPlayer.getRole();
                 if (role == RoleId.Seer || role == RoleId.MadSeer || role == RoleId.EvilSeer)
